Move multilayer error scenarios tests to its own file
